LayoutTests/js/typedarray-slice-expected.txt

This test checks the behavior of the TypedArray.prototype.slice function

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Int32Array.prototype.slice.length is 2
PASS Int32Array.prototype.slice.name is 'slice'
PASS isSameFunctionForEachTypedArrayPrototype('slice') is true
PASS testPrototypeReceivesArray('slice', [undefined, this, { }, [ ], true, '']) is true

testPrototypeFunction has the following arg list (name, args, init, result [ , expectedArray ])

1.0 Test Basic Functionality
PASS testPrototypeFunction('slice', '(2, 3)', [12, 5, 8, 13, 44], [8], [12, 5, 8, 13, 44]) is true
PASS testPrototypeFunction('slice', '(5, 5)', [12, 5, 8, 13, 44], []) is true
PASS testPrototypeFunction('slice', '(0, 5)', [12, 5, 8, 13, 44], [12, 5, 8, 13, 44]) is true
PASS testPrototypeFunction('slice', '(0, -5)', [12, 5, 8, 13, 44], []) is true
PASS testPrototypeFunction('slice', '(-3, -2)', [12, 5, 8, 13, 44], [8]) is true
PASS testPrototypeFunction('slice', '(4, 2)', [12, 5, 8, 13, 44], []) is true
PASS testPrototypeFunction('slice', '(-50, 50)', [12, 5, 8, 13, 44], [12, 5, 8, 13, 44]) is true

2.0 Preserve Underlying bits
PASS intView[0] is -1

3.0 Creates New Buffer
PASS intView[0] is 0

PASS successfullyParsed is true

TEST COMPLETE
